# docs/http-server/apidoc/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Documentation for the exposed HTTP API.",
        "title": "Toolboard API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/dashboard": {
            "put": {
                "description": "Stores the dashboard object into the database with the same ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dashboard"
                ],
                "summary": "Update an existing Dashboard",
                "operationId": "put-dashboard",
                "parameters": [
                    {
                        "description": "The dashboard to update",
                        "name": "dashboard",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Dashboard"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Dashboard"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "post": {
                "description": "Stores the body dashboard object as a new entry in the database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dashboard"
                ],
                "summary": "Create a new Dashboard",
                "operationId": "post-dashboard",
                "parameters": [
                    {
                        "description": "The dashboard to create",
                        "name": "dashboard",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Dashboard"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/Dashboard"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/dashboard/all": {
            "get": {
                "description": "Returns one json object with all dashboards in the database",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dashboard"
                ],
                "summary": "Get all dashboards",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Dashboard"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/dashboard/{id}": {
            "delete": {
                "description": "Checks the ID in the database, and deletes that entry if present",
                "tags": [
                    "dashboard"
                ],
                "summary": "Delete a dashboard",
                "operationId": "delete-dashboard",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Dashboard ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/widget/all": {
            "get": {
                "description": "Returns one json object with all available widgets",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "widget"
                ],
                "summary": "Get all widgets, locally and externally",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Widget"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/widget/embed/{id}/{file}": {
            "get": {
                "description": "Returns html of the Widget with the smae ID",
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "widget"
                ],
                "summary": "Get HTML of a Widget",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Widget ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Optional file name",
                        "name": "file",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "415": {
                        "description": "Unsupported Media Type"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/widget/{id}": {
            "get": {
                "description": "Returns a json object of the Widget with the smae ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "widget"
                ],
                "summary": "Get widget by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Widget ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Widget"
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        }
    },
    "definitions": {
        "Dashboard": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "metadata": {
                    "$ref": "#/definitions/DashboardMetadata"
                },
                "updatedAt": {
                    "type": "string"
                },
                "widgets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/DashboardWidget"
                    }
                }
            }
        },
        "DashboardMetadata": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "dashboardId": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "displayName": {
                    "type": "string"
                },
                "headerImage": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/DashboardTag"
                    }
                },
                "updateAt": {
                    "type": "string"
                }
            }
        },
        "DashboardTag": {
            "type": "object",
            "properties": {
                "bgColor": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "dashboardMetadataId": {
                    "type": "integer"
                },
                "displayName": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "textColor": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "DashboardWidget": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "dashboardId": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "displayName": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "location": {
                    "$ref": "#/definitions/DashboardWidgetLocation"
                },
                "locationId": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                },
                "widget": {
                    "$ref": "#/definitions/Widget"
                },
                "widgetId": {
                    "type": "integer"
                }
            }
        },
        "DashboardWidgetLocation": {
            "type": "object",
            "properties": {
                "MaxWidth": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "height": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "maxHeight": {
                    "type": "integer"
                },
                "minHeight": {
                    "type": "integer"
                },
                "minWidth": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                },
                "width": {
                    "type": "integer"
                },
                "x": {
                    "type": "integer"
                },
                "y": {
                    "type": "integer"
                }
            }
        },
        "Widget": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "installPath": {
                    "type": "string"
                },
                "widgetJSON": {
                    "type": "string"
                }
            }
        }
    }
}